backend/requirements.txt
# Core
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0

# --- Uncomment as needed for your implementation ---

# LLM Providers
# openai>=1.0
# anthropic>=0.30

# RAG / Embeddings
# langchain>=0.2
# langchain-openai>=0.1
# langchain-community>=0.2
# chromadb>=0.5
# pgvector>=0.3
# tiktoken>=0.7

# Async Processing (stretch goal)
# celery>=5.4
# redis>=5.0
# django-celery-results>=2.5
